Add -Werror to SYCL build and fix warnings
- Add -Wall -Wextra -Werror with appropriate suppressions to dpcpp.cmake
- Remove unused hasDriverExtension function and hDriver variable
- Fix const return types on uint32_t/bool functions
- Fix sign comparison in instance_array.h and thread.cpp

+++ common/sys/thread.cpp
@@ -228,7 +228,7 @@ namespace embree
       {
         if (!CPU_ISSET(i,&set)) continue;
 
-        if (j == ID) {
+        if (size_t(j) == ID) {
           ID = i;
           break;
         }
